fixed result stylizing + links
