+++ pages/_app.js
@@ -1,6 +1,6 @@
 import '@/styles/globals.css'
 import { useRouter } from 'next/router'
-import Index from './index'
+import Index from './home'
 
 
 export default function App({ Component, pageProps }) {
